Refactor enemies and level helpers into separate modules
